feat: advanced barrel exports support computedMemberExpr with static key
