+++ crates/rolldown/tests/rolldown/tree_shaking/advanced_barrel_exports2/main.js
@@ -2,3 +2,4 @@ import assert from 'node:assert';
 import * as ns from './shared'
 
 assert.equal(ns.a.b, 500)
+assert.equal(ns.a['a'], 100)
