+++ crates/rolldown/tests/rolldown/tree_shaking/advanced_barrel_exports_bailout_dynamic_key/_config.json
@@ -1,0 +1,1 @@
+{}
