+++ crates/rolldown/tests/rolldown/tree_shaking/advanced_barrel_exports_bailout_dynamic_key/a.js
