+++ crates/rolldown/tests/rolldown/tree_shaking/advanced_barrel_exports_bailout_dynamic_key/c.js
@@ -1,0 +1,1 @@
+export * as c from './a.js'
