+++ crates/rolldown/tests/rolldown/tree_shaking/advanced_barrel_exports_bailout_dynamic_key/main.js
@@ -1,0 +1,6 @@
+import assert from 'node:assert';
+import * as ns from './shared'
+
+let q = 'a'
+// should include all symbols in `a`
+assert.equal(ns.a[q], 100)
