+++ crates/rolldown/tests/rolldown/tree_shaking/advanced_barrel_exports_bailout_dynamic_key/shared.js
@@ -1,0 +1,3 @@
+import {c} from './c.js'
+
+export { c as a }
