+++ scripts/snap-diff/stats/aggregated-reason-without-not-support.md
@@ -165,8 +165,6 @@
 - crates/rolldown/tests/esbuild/loader/loader_file_relative_path_asset_names_css
 ## abs output base
 - crates/rolldown/tests/esbuild/loader/loader_file_relative_path_js
-## json partial namespace memberExpr used tree shaking
-- crates/rolldown/tests/esbuild/loader/loader_json_invalid_identifier_es6
 ## should treated it as cjs module
 - crates/rolldown/tests/esbuild/loader/loader_json_no_bundle
 ## should not transform `export * as ns from 'mod'` above es2019
