+++ scripts/snap-diff/summary/loader.md
@@ -65,8 +65,6 @@
   missing
 ## loader_from_extension_with_query_parameter
   missing
-## [loader_json_invalid_identifier_es6](../../../crates/rolldown/tests/esbuild/loader/loader_json_invalid_identifier_es6/diff.md)
-  diff
 ## [loader_json_no_bundle](../../../crates/rolldown/tests/esbuild/loader/loader_json_no_bundle/diff.md)
   diff
 ## [loader_json_no_bundle_common_js](../../../crates/rolldown/tests/esbuild/loader/loader_json_no_bundle_common_js/diff.md)
@@ -102,6 +100,7 @@
 ## [loader_data_url_text_java_script_plus_character](../../../crates/rolldown/tests/esbuild/loader/loader_data_url_text_java_script_plus_character/bypass.md)
 ## [loader_file_common_js_and_es6](../../../crates/rolldown/tests/esbuild/loader/loader_file_common_js_and_es6/bypass.md)
 ## [loader_json_common_js_and_es6](../../../crates/rolldown/tests/esbuild/loader/loader_json_common_js_and_es6/bypass.md)
+## [loader_json_invalid_identifier_es6](../../../crates/rolldown/tests/esbuild/loader/loader_json_invalid_identifier_es6/bypass.md)
 ## [loader_json_no_bundle_es6](../../../crates/rolldown/tests/esbuild/loader/loader_json_no_bundle_es6/bypass.md)
 ## [loader_json_prototype](../../../crates/rolldown/tests/esbuild/loader/loader_json_prototype/bypass.md)
 ## [loader_json_prototype_es5](../../../crates/rolldown/tests/esbuild/loader/loader_json_prototype_es5/bypass.md)
